% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepath.R
\name{deepath}
\alias{deepath}
\title{omics enrichment analysis}
\usage{
deepath(
  input_data,
  output,
  mapper_file,
  pathway_col = "Pathway",
  feature_col = "Feature",
  input_metadata = NA,
  meta = NA,
  case_label = NA,
  control_label = NA,
  score_col = "logFC",
  pval_threshold = 0.05,
  fdr_threshold = NA,
  Pathway.Subject = NA,
  method = "ks",
  min_member = 2,
  do_plot = TRUE
)
}
\arguments{
\item{input_data}{It could be either 1) a score file which has at least a column for feature names and an column for scores, or 2) a omics profile which has rows as samples and columns as features}

\item{output}{output a directory path to where user wants resulst to be written}

\item{mapper_file}{a mapping file that has at least towcolumns one for pathways and one for features in each pathways in long format}

\item{pathway_col}{a column for pathways in the mapper_file}

\item{feature_col}{a column for features in the mapper_file}

\item{input_metadata}{a file or data frame that has metadata where score file in not provided and scores need to be calulated usig input_data, input_metadata, and meta}

\item{meta}{a couln name in input_metadata when score needs to be calculated an d score file is not provided (input_data is not a core file but a omics profile)}

\item{case_label}{a name for case group and is part of meta values.}

\item{control_label}{a name for control group and is part of meta values.}

\item{score_col}{a columnb name in input_data}

\item{pval_threshold}{a threshold for p-value which will be used to decide to visualize result for a pathway or not}

\item{fdr_threshold}{a threshold for fdr which will be used to decide to visualize result for a pathway or not}

\item{Pathway.Subject}{a pathway subject when HMDB database is provided as reference for metabolites (as omics features)}

\item{method}{a enrichment methods to be used, options are ks, gset, and wilcox}

\item{min_member}{a condition for a pathway to be analyzed as minimum number of a pathway members to be seen in the study}

\item{do_plot}{a TRUE or FALSE to plot result or not}
}
\value{
The result as a list  of 1) \code{enrichment_stats} a table for statistic of pathway enrichment, \code{rank_plots} a list of ggplots for density plot of enrichment base don rank of features, and 3) \code{score_plots} a list of ggplots for density plot of enrichment base don score of feature.
}
\description{
omics enrichment analysis
}
\examples{
deepath_result <- deepath::deepath( input_data = score_data_no_COVID, input_metadata = NA, 
pathway_col = "Pathway", feature_col = "Feature", meta <- NA, case_label <- "", 
control_label <- "", output = "~/deepath_enrichment_metabolite_no_COVID_fdr1",
input_data = score_data_no_COVID,
score_col = 'coef', 
pval_threshold = 0.05, 
fdr_threshold = NA, 
Pathway.Subject = NA,#'Metabolic', 
do_plot = TRUE, 
mapper_file = "~/Databases/Metabolomics/Pathways/smpdb_metabolites.tsv",
method = "wilcox",
min_member = 2)
}
