% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GPD Permutation Test}
\alias{GPD Permutation Test}
\alias{gpd_permutation_test}
\title{Implementation of the GPD-based p-value estimation algorithm}
\usage{
gpd_permutation_test(
  x0,
  y,
  minM_epdf = 10,
  Nexc = 250,
  Nexc_shrink = 10,
  Nexc_alpha = 0.05,
  yfun,
  ystart = 200,
  ygrow = 100,
  ymax = 5000
)
}
\arguments{
\item{x0}{an input vector of values}

\item{y}{an input vector of values}

\item{minM_epdf}{a minimumn number of iterations for EPDF}

\item{Nexc}{a number of permution to performe in each step default 250}

\item{Nexc_shrink}{default 10}

\item{Nexc_alpha}{default 0.05}

\item{yfun}{default}

\item{ystart}{default 200}

\item{ygrow}{default 100}

\item{ymax}{default 5000}
}
\value{
a `pvalue`
}
\description{
Implementation of the GPD-based p-value estimation algorithm
}
\examples{
gpd_permutation_test(x, y)
}
\references{
Implementation of the GPD-based p-value estimation algorithm from
Knijnenburg, Wessels, Reinders, and Shmulevich (2009) Fewer
permutations, more accurate P-values. Bioinformatics 25(12): i161–i168.
DOI: 10.1093/bioinformatics/btp211
Implemented by Jason Lloyd-Price and Ali Rahnavard
}
